Create README.md
Fake news is a type of yellow journalism that incorporates pieces of news that may be hoaxes and is widely disseminated via social media and other internet platforms. This is frequently done to promote or impose specific views, and it is frequently accomplished through political agendas. Such news pieces may contain misleading and/or inflated claims, and algorithms may spread them, trapping users in a filter bubble.